--- controllers/file.go
@@ -1,3 +1,0 @@
-package controllers
-
-type FileController struct{}
