--- controllers/package.go
@@ -1,17 +1,0 @@
-package controllers
-
-type PackageController struct{}
-
-//// CreateRootAndNestedFolders creates the collection packages and/or returns existing folders
-//func (*PackageController) CreateRootAndNestedFolders(rootFolder pkg.FolderUpload, files []pkg.UploadFile) []models.Package {
-//
-//	//var uploadCollections []pkg.FolderUpload
-//	//for f := range files {
-//	//
-//	//}
-//
-//}
-
-type Container struct {
-	organizationId int
-}
