+++ lambda/authorizer/handler/handler.go
@@ -9,10 +9,10 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/dynamodb"
 	"github.com/lestrrat-go/jwx/v2/jwk"
 	"github.com/lestrrat-go/jwx/v2/jwt"
-	"github.com/pennsieve/pennsieve-go-api/models/dataset"
-	"github.com/pennsieve/pennsieve-go-api/models/dbTable"
 	"github.com/pennsieve/pennsieve-go-api/pkg/authorizer"
 	"github.com/pennsieve/pennsieve-go-api/pkg/core"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/dataset"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/dbTable"
 	"log"
 	"os"
 	"regexp"
