+++ pkg/models/dbTable/fileTable.go
@@ -5,11 +5,11 @@ import (
 	"fmt"
 	"github.com/google/uuid"
 	"github.com/lib/pq"
-	"github.com/pennsieve/pennsieve-go-api/models/fileInfo/fileType"
-	"github.com/pennsieve/pennsieve-go-api/models/fileInfo/objectType"
-	"github.com/pennsieve/pennsieve-go-api/models/fileInfo/processingState"
-	"github.com/pennsieve/pennsieve-go-api/models/fileInfo/uploadState"
 	"github.com/pennsieve/pennsieve-go-api/pkg/core"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/fileInfo/fileType"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/fileInfo/objectType"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/fileInfo/processingState"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/fileInfo/uploadState"
 	"log"
 	"strings"
 	"time"
