+++ pkg/models/dbTable/packageTable.go
@@ -5,10 +5,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/lib/pq"
-	"github.com/pennsieve/pennsieve-go-api/models/packageInfo"
-	"github.com/pennsieve/pennsieve-go-api/models/packageInfo/packageState"
-	"github.com/pennsieve/pennsieve-go-api/models/packageInfo/packageType"
 	"github.com/pennsieve/pennsieve-go-api/pkg/core"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/packageInfo"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/packageInfo/packageState"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/packageInfo/packageType"
 	"log"
 	"regexp"
 	"strings"
