+++ pkg/models/organization/models.go
@@ -2,14 +2,14 @@ package organization
 
 import (
 	"fmt"
-	"github.com/pennsieve/pennsieve-go-api/models/dbTable"
+	dbTable2 "github.com/pennsieve/pennsieve-go-api/pkg/models/dbTable"
 )
 
 // Claim combines the role of the user in the org, and the features in the organization.
 type Claim struct {
-	Role            dbTable.DbPermission
+	Role            dbTable2.DbPermission
 	IntId           int64
-	EnabledFeatures []dbTable.FeatureFlags
+	EnabledFeatures []dbTable2.FeatureFlags
 }
 
 func (c Claim) String() string {
