+++ pkg/models/packageInfo/packageAttributes.go
