+++ pkg/models/uploadFile/upload_file.go
@@ -4,10 +4,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/google/uuid"
-	"github.com/pennsieve/pennsieve-go-api/models/fileInfo/fileType"
-	"github.com/pennsieve/pennsieve-go-api/models/iconInfo"
-	"github.com/pennsieve/pennsieve-go-api/models/packageInfo/packageType"
-	"github.com/pennsieve/pennsieve-go-api/models/uploadFolder"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/fileInfo/fileType"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/iconInfo"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/packageInfo/packageType"
+	"github.com/pennsieve/pennsieve-go-api/pkg/models/uploadFolder"
 	"sort"
 	"strings"
 )
